refactor: 설정 항목을 컴포넌트 단위로 세분화(#29)
