+++ src/components/commons/SettingModal/AccountDeletionSection.tsx
@@ -1,0 +1,3 @@
+export function AccountDeletionSection() {
+  return <div className="py-6">회원 탈퇴</div>;
+}
